Redefine Special Skill, Item Description
특수기 구조체 변경: Value1T 타입을 float에서 int32로 변경(처음 구조체를 생성할 때 타입을 잘못 선언하여 수정함)
특수기, 아이템, 페피 턴에도 스킬 설명 중 '{번호}'를 찾고 그 값을 구하여 스킬 설명을 재정의 후 적용

+++ Source/KNOCKturne/BattleTableManagerSystem.h
@@ -241,7 +241,7 @@ struct FSpecialSkillData : public FTableRowBase {
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
 	float value1M;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
-	float value1T;
+	int32 value1T;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
 	FString BuffCode;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
